修复：根据图书馆ID查询最近藏书量、处理图书借阅、处理管理员确认图书归还 三个控制器方法已经更新；读者借阅-管理员审批借阅-读者归还-管理员确认归还 流程得到更新优化（图书是否已归还逻辑使用库存表判断）；修复了前端的Bug；新增book_order表图书订购表，现在可以追踪每笔图书的订购；更新了借阅表中部分数据的借阅方式字段；待确认：图书表中的冗余字段移除出表结构后的影响；未完成：图书订购时尚不能同步更新图书订购表

+++ ruoyi-admin/src/main/java/com/ruoyi/book/mapper/BooksMapper.java
@@ -69,7 +69,7 @@ public interface BooksMapper
      * @param ids
      * @return
      */
-    public List<Books> selectBooksByIds(List<Integer> ids);
+    public List<Books> selectBooksByIds(List<Long> ids);
 
 
 }
